refactor(vsock-guest): split exec connection tests
