+++ crates/vsock-guest/tests/connection/main.rs
@@ -8,7 +8,13 @@
 mod support;
 
 mod basic;
-mod exec_operation;
+mod exec_basic;
+mod exec_cancel_cleanup;
+mod exec_control;
+mod exec_helpers;
+mod exec_output;
+mod exec_stdin;
+mod exec_validation;
 mod quiesce;
 mod reconnect;
 mod shutdown;
